fix:Refactored Steps and Sidecars to container_validation #7454
